buh-bot/requirements.txt
maxapi==0.1.7
apscheduler==3.10.4
python-dotenv==1.0.1
pytz==2024.1
openpyxl==3.1.5
